REFACTOR: big hatchet through the binding+package files, splitting into sub-api modules
